Dark Elf Light Sensitivity + Dark Vision (#2263)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

This is the first of a planned series of PRs to add more unique
mechanics to various races. The goal is to create a more immersive
experience to allow players to really feel how a given race is unique
from those around them. I'll be starting with the oft-maligned "Shunned"
races with hopes of giving them mechanics that are unique, immersive,
and help to drive a wedge between them and the rest of Rockhill.

What does it do?

- Dark Elves now have the "Darkling" trait. This trait grants them
darkvision, but also makes them sensitive to light.
- Dark Elves accrue light stress when exposed to bright light. Their
resistance is based on their endurance stat. (Note, light stress is NOT
the kind of stress that gives you heart attacks)
- Once a dark elf has accrued enough light stress, they receive a debuff
to perception, a stress debuff and a slightly blurred screen.
- If a dark elf continues to accrue light stress they will begin to have
migraines, blurring their vision further and debuffing their stats.
- Resting in the dark or with their eyes closed reduces light stress.
Sleeping reduces it even faster.
- Being exposed to very bright lights will momentarily blind the dark
elf. Resilience is determined by endurance stat. This effect has a
cooldown which is determined by the dark elf's constitution score.
- Wearing a hood or mask, anything that restricts vision, increases
resistance to light.
- Direct exposure to daylight counts for double!

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

More immersive, unique mechanics to start making our races feel more
interesting than just a statblock.

Dark elves are now incentivized to stick to the shadows, hide in dark
places and go around with their faces covered. This should help to make
them feel more shunned and ostracized, and to make the town feel like a
hostile place that isn't made for their people. Mask laws, overzealous
guards and the occasional misunderstanding will work against dark elves
trying to live among these strange, light-loving surface dwellers.

Meanwhile, dark elves are comfortable at home underground and stalking
at night. As they should be.

Pay no heed to the rous AI bugging out in the background, that's not me.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

![image](https://github.com/user-attachments/assets/8a843386-069b-403c-a00b-00581e85d23f)

![image](https://github.com/user-attachments/assets/2e88397e-0a00-464c-9059-1df9bcd4a758)

![image](https://github.com/user-attachments/assets/448208e9-fd3f-456f-bee1-c86800a3a387)

![image](https://github.com/user-attachments/assets/55987170-feed-4ae7-b970-ec5d830da204)

![image](https://github.com/user-attachments/assets/e000649b-3623-44f7-a1b4-56d6d3940a15)

![image](https://github.com/user-attachments/assets/6d7e3a67-83a9-4e19-b50d-d8ef5f2812fa)

![image](https://github.com/user-attachments/assets/39e98442-250e-48c9-912e-2c655c3e9c83)

![image](https://github.com/user-attachments/assets/e424705d-080e-4470-8cba-9dc7aed2fc2e)

Video demonstration, sorry about the dogwater compression.

https://github.com/user-attachments/assets/baac1435-3d2e-4d73-92d3-d167506c1ebf

---------

Co-authored-by: Spaghetti-bit <yumyumkillkill@gmail.com>